refactor(tests): extract .fptest parser into noir_ieee754_inputs package
Move the IBM FPgen .fptest grammar (parse_test_line, parse_fp_value,
parse_fptest_file, KNOWN_BAD_TESTS) plus the FPValue / TestCase /
Operation / Precision / RoundingMode dataclasses out of
scripts/generate_tests.py into a colocated package
scripts/noir_ieee754_inputs/ with a dedicated fptest submodule.

fp_value_to_bits stays in fptest.py for now; phase 2c rewrites it on
top of float.fromhex + struct.

Generator output is byte-identical: regen of test_packages and the
CI matrix produces no diff against the baseline.
